requirements.txt
pandas
numpy
pathlib
pytest
regex
argparse
datetime
xlsxwriter
google-cloud-storage
google-cloud-bigquery
google-api-core>=2.0.0

